Add files via upload
Nth prime file is to find the prime nth prime number of a given in integer.
distancetoprimes is to find the distance between the first prime number above a given in integer and first prime number below a given in integer. 
i.e 10 first above is 11 , first prime below is 7 , 11-7 = 4 distance between is 4.